CLN: rename script filename
