+++ geopandas/_decorator.py
